Ordered content with lists
Still needs work, maybe you could figure out how to make divider a bit longer, use Barcelona sigil as an inline emoji. Also block of Titles Won needs more thought regarding where to put it now that it is narrow.
P.S. Usually it is better to split a commit like this one with lot of changes into several smaller, but I was in a hurry and also got driven away a bit, so here is everything together, that means if you don't like some changes it is harder to revert, I'm sorry.